intermediate working version

+++ src/main/java/com/example/demo/service/PeopleService.java
@@ -5,8 +5,10 @@ import com.example.demo.errors.PersonNotDeletedException;
 import com.example.demo.models.Person;
 import com.example.demo.repository.person.PersonRepository;
 import com.example.demo.utils.mappers.PeopleMapper;
+import lombok.AccessLevel;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -19,10 +21,17 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
 public class PeopleService {
 
     final PersonRepository personRepository;
     final PeopleMapper peopleMapper;
+
+    public String getNextPersonMediaId() {
+        Long maxId = personRepository.findMaxPersonId().orElse(0L);
+        Long nextId = maxId + 1;
+        return String.format("EDC%08d", nextId);
+    }
 
     public List<PersonResponse> getAllPeople() {
         List<Person> allPeople = personRepository.findAll();
@@ -36,14 +45,16 @@ public class PeopleService {
         return new PageImpl<>(responses, pageable, personPage.getTotalElements());
     }
 
-
     public Optional<Person> getPersonById(@NonNull Long person_id) {
-        return personRepository.findById(person_id);
+        return personRepository.findByPersonId(person_id);
     }
 
     public void save(@NonNull Person person) {
+        final String nextPersonMediaId = this.getNextPersonMediaId();
+        person.setPersonMediaId(nextPersonMediaId);
+
         personRepository.save(person);
-        System.out.println("PeopleService.class save person to DB personId = " + person.getPerson_id());
+        System.out.println("PeopleService.class save person to DB personId = " + person.getPersonId());
     }
 
     public void update(@NonNull Integer id, @NonNull Person updatedPerson) {
@@ -51,9 +62,9 @@ public class PeopleService {
     }
 
     public void delete(@NonNull Long id) {
-        Optional<Person> personById = personRepository.findById(id);
+        Optional<Person> personById = personRepository.findByPersonId(id);
 
-        if (personById.isPresent()) {
+        if (personById.isPresent() && personById.get().getBooks().isEmpty()) {
             personRepository.delete(id);
         } else {
             throw new PersonNotDeletedException(id);
@@ -78,7 +89,7 @@ public class PeopleService {
 
         long start = System.currentTimeMillis();
 
-        personRepository.butchSaveAll(people); // ❗ убедись, что репозиторий поддерживает saveAll
+        personRepository.butchSaveAll(people);
 
         long end = System.currentTimeMillis();
         System.out.println("⏱ Batch вставка 1000 записей заняла: " + (end - start) + " мс");
@@ -92,12 +103,19 @@ public class PeopleService {
             int newLastPersonId = lastPersonIdFromDB + 1;
             int newLastPersonIdPlus1000 = lastPersonIdFromDB + 1001;
 
+            Long maxId = personRepository.findMaxPersonId().orElse(0L);
+
             for (int i = newLastPersonId; i < newLastPersonIdPlus1000; i++) {
-                people.add(new Person("persone_" + i, 10, "persone_" + i + "@gmail.com", "city_" + i));
+                Long nextId = ++maxId;
+
+                people.add(new Person(String.format("EDC%08d", nextId), "persone_" + i, 10, "persone_" + i + "@gmail.com", "city_" + i));
             }
         } else {
+            Long maxId = personRepository.findMaxPersonId().orElse(0L);
+
             for (int i = 1; i < 1001; i++) {
-                people.add(new Person("persone_" + i, 10, "persone_" + i + "@gmail.com", "city_" + i));
+                Long nextId = ++maxId;
+                people.add(new Person(String.format("EDC%08d", nextId), "persone_" + i, 10, "persone_" + i + "@gmail.com", "city_" + i));
             }
         }
         return people;
